+++ src/TestFramework/TestFramework.Extensions/TestFramework.Extensions.csproj
@@ -82,6 +82,10 @@
   </ItemGroup>
 
   <ItemGroup>
+    <InternalsVisibleTo Include="Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter" Key="$(VsPublicKey)" />
+  </ItemGroup>
+
+  <ItemGroup>
     <None Update="build\winui\MSTest.TestFramework.targets">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
