+++ test/IntegrationTests/MSTest.Acceptance.IntegrationTests/Properties/launchSettings.json
@@ -2,7 +2,7 @@
   "profiles": {
     "MSTest.Acceptance.IntegrationTests": {
       "commandName": "Project",
-      "commandLineArgs": ""
+      //"commandLineArgs": "--filter ClassName~TimeoutTests"
     }
   }
 }
